Refactor serializers module into package
The serializers module has grown so big that it makes sense to split
it into a package.

Refs LIIK-183
