+++ traffic_control/serializers/road_marking.py
@@ -1,0 +1,74 @@
+from enumfields.drf import EnumSupportSerializerMixin
+from rest_framework import serializers
+from rest_framework_gis.fields import GeometryField
+
+from ..models import (
+    RoadMarkingPlan,
+    RoadMarkingPlanFile,
+    RoadMarkingReal,
+    RoadMarkingRealFile,
+    TrafficControlDeviceType,
+)
+from ..models.common import DeviceTypeTargetModel
+
+
+class RoadMarkingPlanFileSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = RoadMarkingPlanFile
+        fields = "__all__"
+
+
+class RoadMarkingPlanSerializer(
+    EnumSupportSerializerMixin, serializers.ModelSerializer
+):
+    files = RoadMarkingPlanFileSerializer(many=True, read_only=True)
+    device_type = serializers.PrimaryKeyRelatedField(
+        queryset=TrafficControlDeviceType.objects.for_target_model(
+            DeviceTypeTargetModel.ROAD_MARKING
+        )
+    )
+
+    class Meta:
+        model = RoadMarkingPlan
+        read_only_fields = (
+            "created_by",
+            "updated_by",
+            "deleted_by",
+            "deleted_at",
+        )
+        exclude = ("is_active", "deleted_at", "deleted_by")
+
+
+class RoadMarkingPlanGeoJSONSerializer(RoadMarkingPlanSerializer):
+    location = GeometryField()
+
+
+class RoadMarkingRealFileSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = RoadMarkingRealFile
+        fields = "__all__"
+
+
+class RoadMarkingRealSerializer(
+    EnumSupportSerializerMixin, serializers.ModelSerializer
+):
+    files = RoadMarkingRealFileSerializer(many=True, read_only=True)
+    device_type = serializers.PrimaryKeyRelatedField(
+        queryset=TrafficControlDeviceType.objects.for_target_model(
+            DeviceTypeTargetModel.ROAD_MARKING
+        )
+    )
+
+    class Meta:
+        model = RoadMarkingReal
+        read_only_fields = (
+            "created_by",
+            "updated_by",
+            "deleted_by",
+            "deleted_at",
+        )
+        exclude = ("is_active", "deleted_at", "deleted_by")
+
+
+class RoadMarkingRealGeoJSONSerializer(RoadMarkingRealSerializer):
+    location = GeometryField()
