+++ traffic_control/serializers/traffic_light.py
@@ -1,0 +1,74 @@
+from enumfields.drf import EnumSupportSerializerMixin
+from rest_framework import serializers
+from rest_framework_gis.fields import GeometryField
+
+from ..models import (
+    TrafficControlDeviceType,
+    TrafficLightPlan,
+    TrafficLightPlanFile,
+    TrafficLightReal,
+    TrafficLightRealFile,
+)
+from ..models.common import DeviceTypeTargetModel
+
+
+class TrafficLightPlanFileSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = TrafficLightPlanFile
+        fields = "__all__"
+
+
+class TrafficLightPlanSerializer(
+    EnumSupportSerializerMixin, serializers.ModelSerializer
+):
+    files = TrafficLightPlanFileSerializer(many=True, read_only=True)
+    device_type = serializers.PrimaryKeyRelatedField(
+        queryset=TrafficControlDeviceType.objects.for_target_model(
+            DeviceTypeTargetModel.TRAFFIC_LIGHT
+        )
+    )
+
+    class Meta:
+        model = TrafficLightPlan
+        read_only_fields = (
+            "created_by",
+            "updated_by",
+            "deleted_by",
+            "deleted_at",
+        )
+        exclude = ("mount_type", "is_active", "deleted_at", "deleted_by")
+
+
+class TrafficLightPlanGeoJSONSerializer(TrafficLightPlanSerializer):
+    location = GeometryField()
+
+
+class TrafficLightRealFileSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = TrafficLightRealFile
+        fields = "__all__"
+
+
+class TrafficLightRealSerializer(
+    EnumSupportSerializerMixin, serializers.ModelSerializer
+):
+    files = TrafficLightRealFileSerializer(many=True, read_only=True)
+    device_type = serializers.PrimaryKeyRelatedField(
+        queryset=TrafficControlDeviceType.objects.for_target_model(
+            DeviceTypeTargetModel.TRAFFIC_LIGHT
+        )
+    )
+
+    class Meta:
+        model = TrafficLightReal
+        read_only_fields = (
+            "created_by",
+            "updated_by",
+            "deleted_by",
+            "deleted_at",
+        )
+        exclude = ("mount_type", "is_active", "deleted_at", "deleted_by")
+
+
+class TrafficLightRealGeoJSONSerializer(TrafficLightRealSerializer):
+    location = GeometryField()
