+++ traffic_control/serializers/traffic_sign.py
@@ -1,0 +1,76 @@
+from enumfields.drf import EnumSupportSerializerMixin
+from rest_framework import serializers
+from rest_framework_gis.fields import GeometryField
+
+from ..models import (
+    TrafficControlDeviceType,
+    TrafficSignPlan,
+    TrafficSignPlanFile,
+    TrafficSignReal,
+    TrafficSignRealFile,
+)
+from ..models.common import DeviceTypeTargetModel
+
+
+class TrafficSignPlanFileSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = TrafficSignPlanFile
+        fields = "__all__"
+
+
+class TrafficSignPlanSerializer(
+    EnumSupportSerializerMixin, serializers.ModelSerializer
+):
+    files = TrafficSignPlanFileSerializer(many=True, read_only=True)
+    device_type = serializers.PrimaryKeyRelatedField(
+        queryset=TrafficControlDeviceType.objects.for_target_model(
+            DeviceTypeTargetModel.TRAFFIC_SIGN
+        )
+    )
+
+    class Meta:
+        model = TrafficSignPlan
+        read_only_fields = (
+            "created_by",
+            "updated_by",
+            "deleted_by",
+            "deleted_at",
+        )
+        exclude = ("mount_type", "is_active", "deleted_at", "deleted_by")
+
+
+class TrafficSignPlanGeoJSONSerializer(TrafficSignPlanSerializer):
+    location = GeometryField()
+    affect_area = GeometryField(required=False)
+
+
+class TrafficSignRealFileSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = TrafficSignRealFile
+        fields = "__all__"
+
+
+class TrafficSignRealSerializer(
+    EnumSupportSerializerMixin, serializers.ModelSerializer
+):
+    files = TrafficSignRealFileSerializer(many=True, read_only=True)
+    device_type = serializers.PrimaryKeyRelatedField(
+        queryset=TrafficControlDeviceType.objects.for_target_model(
+            DeviceTypeTargetModel.TRAFFIC_SIGN
+        )
+    )
+
+    class Meta:
+        model = TrafficSignReal
+        read_only_fields = (
+            "created_by",
+            "updated_by",
+            "deleted_by",
+            "deleted_at",
+        )
+        exclude = ("mount_type", "legacy_code", "is_active", "deleted_at", "deleted_by")
+
+
+class TrafficSignRealGeoJSONSerializer(TrafficSignRealSerializer):
+    location = GeometryField()
+    affect_area = GeometryField(required=False)
