+++ traffic_control/views/barrier.py
@@ -11,7 +11,7 @@ from ..schema import (
     location_parameter,
     MultiFileUploadSchema,
 )
-from ..serializers import (
+from ..serializers.barrier import (
     BarrierPlanFileSerializer,
     BarrierPlanGeoJSONSerializer,
     BarrierPlanSerializer,
