+++ traffic_control/views/signpost.py
@@ -11,7 +11,7 @@ from ..schema import (
     location_parameter,
     MultiFileUploadSchema,
 )
-from ..serializers import (
+from ..serializers.signpost import (
     SignpostPlanFileSerializer,
     SignpostPlanGeoJSONSerializer,
     SignpostPlanSerializer,
